Add files via upload
This project is a responsive grocery website built using HTML, CSS, and Bootstrap. The website provides a user-friendly interface for browsing and purchasing grocery items. It features well-organized sections for product categories, a search bar, and an easy-to-navigate layout. The site is designed to be fully responsive, ensuring it works seamlessly on mobile, tablet, and desktop devices. Bootstrap's grid system and components were utilized to create a clean, modern design with smooth navigation and a focus on accessibility and user experience.